+++ test/pool_test.dart
@@ -697,6 +697,11 @@ void main() {
       });
     });
   });
+
+  test("throw error when pool limit <= 0", () {
+    expect(() => Pool(-1), throwsArgumentError);
+    expect(() => Pool(0), throwsArgumentError);
+  });
 }
 
 /// Returns a function that will cause the test to fail if it's called.